Zefix gives different results
It looks like Zefix has been updated from the previous version. This break the
analysis done. However, the analysis should be re-runned soon since it's
possible that this issue is due to Zefix heavy used.

Old cache files were stored on `cache/old` as a copy. However, they're not
consistent in the sense that the most recent files have the same issues from
Zefix.